third_party/WebKit/LayoutTests/platform/win/external/wpt/html/semantics/forms/the-input-element/checked-expected.txt

This is a testharness.js-based test.
PASS input@checked is immediately reflected to 'checked' IDL attribute 
Harness: the test ran to completion.
